chore: fix config initialization issue
